fix: split services rfid and pedometer
